move installer to main module
Installer can now be once again be invoked with `python -m
fancycompleter install`
